extract cache logic to separated module
